MJ-2175: Save esignature metadata to database
The purpose of the work contained here is to save the metadata (ID,
    CompanyID, and UploadDate among other properties) of created
templates and signature requests into a common table so that pagination
can be more efficient for endpoints that return consolidated data.
- The createTemplate endpoint has been updated to include the uploadDate
and uploadedBy properties on the template's metadata.
- The createBulkSignatureRequest endpoint has been updated to include
the uploadDate property on the request's metadata.
- The createBulkSignatureRequest endpoint has also been updated to
insert a new record containing the request's metadata into the
EsignatureMetadata table on successful creation of the request.
- A new endpoint was created to save a created template's metadata into
the EsignatureMetadata table. The endpoint will also only update the
title of an edited template.