fix: allow large legacy docs to be previewed, downloaded and deleted
- The queryExecutor lambda did not have enough CPU to handle processing
large documents. The CPU was increased in the serverless.yml.
- Additionally, the queryExecutor lambda could not return large
documents back to the invoking lambda. In order to bypass this issue,
the queryExecutor first saves the document to S3 and then returns an S3
object key to the invoking lambda. The invoking lambda then uses the key
to generate a signed URL and returns that to the client. The preview and
delete document endpoints were updated to do this. Also, these endpoints
were updated to check the database first for documents that have already
been saved to S3.