fix: allow batch signature request creation
- There was a bug with the bulk signature request endpoint where it was
adding multiple signers to one signature request, rather than creating
signature requests for multiple employees. The code was refactored
heavily to fix this.
- The bulk signature request endpoint was renamed to batch to better
match its intended functionality, and to be more consistent with
HelloSign naming convention.
- The calls to HelloSign that create the signature requests have been
parallelized to improve performance.
- The endpoint now accepts a special string that allows for the creation
of signature requests for all employees under a company. Eventually, we
will implement a solution that will eliminate the need for this hacky
solution.