refactor: Update companyUpdate and associated functions to take companyId
Currently companyUpdate takes EvoCompanyCode for migration, updated
companyUpdate to take companyId, get AHR Company, and pass the
CompanyCode to associated functions.  Inside those functions, we do
the same thing with the passed value which now includes the new AHR
company ID instead of new company EvoCompanyCode. That is translated
and passed as usual through the functions.